pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lotsearch-engine"
version = "0.1.0"
description = "Extract data from Lotsearch environmental PDF reports and generate PSI report sections"
requires-python = ">=3.11"
dependencies = [
    "pdfplumber>=0.11.0",
    "pydantic>=2.5.0",
    "jinja2>=3.1.0",
    "click>=8.1.0",
    "anthropic>=0.40.0",
    "python-docx>=1.1.0",
    "rich>=13.0.0",
    "python-dateutil>=2.8.0",
    "streamlit>=1.39.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "ruff",
]

[project.scripts]
lotsearch = "src.cli:main"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
